+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [MKHH-#5] Plik json do textów statycznych, plik ze stałymi dla sass
 - [MKHH-#13] Strona logowania -- Stworzenie komponentów reużywalnych: button
 - [MKHH-#14] Strona logowania -- Projekt layoutu strony
+- [MKHH-#20] Widok użytkownika -- Projekt layoutu strony